refactor: layout & shared 디렉토리 정리
